requirements.txt
absl-py==0.11.0
albumentations==0.5.2
attr==0.3.1
attrs==20.3.0
backcall==0.2.0
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
coremltools==4.0
cycler==0.10.0
dacite==1.6.0
dataclasses==0.6
debugpy==1.5.1
decorator==4.4.2
entrypoints==0.3
fsspec==0.8.4
future==0.18.2
google-auth==1.23.0
google-auth-oauthlib==0.4.2
grpcio==1.34.0
idna==2.10
imageio==2.9.0
imgaug==0.4.0
ipykernel==6.6.0
ipython==7.30.1
jedi==0.18.1
joblib==0.17.0
jupyter-client==7.1.0
jupyter-core==4.9.1
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
matplotlib-inline==0.1.3
mpmath==1.1.0
nest-asyncio==1.5.4
networkx==2.5
numpy==1.19.4
oauthlib==3.1.0
omegaconf==2.0.5
onnx==1.8.0
onnx-coreml==1.3
onnx-simplifier==0.2.19
onnxoptimizer==0.1.1
onnxruntime==1.5.2
opencv-python==4.4.0.46
opencv-python-headless==4.4.0.46
packaging==20.7
pandas==1.1.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pkg_resources==0.0.0
prompt-toolkit==3.0.23
protobuf==3.14.0
ptyprocess==0.7.0
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.10.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-lightning==1.0.8
pytorch-ranger==0.1.1
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==22.3.0
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
Shapely==1.7.1
six==1.15.0
sympy==1.7
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
threadpoolctl==2.1.0
tifffile==2020.11.26
timm==0.5.4
torch==1.9.0+cu111
torch-optimizer==0.0.1a17
torchaudio==0.9.0
torchvision==0.10.0+cu111
tornado==6.1
tqdm==4.54.0
traitlets==5.1.1
typing==3.7.4.3
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
Werkzeug==1.0.1
